GCS FS: Configuration parameters (#1571)
* GCS FS: Complete multi-part uploads

This adds support for multi-part uploads with more than 32 parts. There is
a 32-object limitation for a single ComposeObject() function, but the GCS
provides a helper routine, ComposeMany(), that handles composing more than
32 parts. This will result in multiple requests to GCS.

This also uses another helper function, DeleteByPrefix(), to garbage collect
intermediate composed objects in the event of a failure/crash.

Unit tests have been added to test all the obvious scenarios.

* GCS FS: Configuration parameters

This completes the functional implementation of the GCS FS. The next patches
will be related to CI, testing, and integrating a GCS emulator.